add fluent validation to category entity

+++ Application/Features/Category/Handlers/Commands/DeleteCategoryCommandHandler.cs
@@ -1,24 +1,35 @@
-﻿using Application.Features.Category.Requests.Commands;
-using Application.Persistence.Contracts;
-using AutoMapper;
+﻿using Application.DTOs.Category;
 using MediatR;
+using Application.Exceptions;
+using Application.Persistence.Contracts;
+using Application.DTOs.Category.Validators;
+using Application.Features.Category.Requests.Commands;
 
 namespace Application.Features.Category.Handlers.Commands;
 
-public class DeleteCategoryCommandHandler :IRequestHandler<DeleteCategoryCommand>
+public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand>
 {
     private readonly ICategoryRepository _categoryRepository;
-    private readonly IMapper _mapper;
 
-    public DeleteCategoryCommandHandler(ICategoryRepository categoryRepository, IMapper mapper)
+    public DeleteCategoryCommandHandler(ICategoryRepository categoryRepository)
     {
         _categoryRepository = categoryRepository;
-        _mapper = mapper;
     }
 
     public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
     {
+        var validator = new DeleteCategoryValidator();
+        var validationResult = await validator
+            .ValidateAsync(request.DeleteCategory, cancellationToken);
+        if (!validationResult.IsValid)
+            throw new ValidationException(validationResult);
+
         var category = await _categoryRepository.GetByIdAsync(request.DeleteCategory.Id);
+
+        if (category is null)
+            throw new NotFoundException(nameof(Domain.Shop.Category),request.DeleteCategory.Id);
+        
+        
         await _categoryRepository.DeleteCategoryAsync(category);
         return Unit.Value;
     }
